Remove machine-local Cargo config and update persistence finding invariants
This commit removes `.cargo/config.toml`, which contained a
machine-specific
BOOST_ROOT path that should not be versioned, and updates the
`PersistenceFinding`
trait to allow zero-length spans.

The trait invariant is relaxed from `span_end() > span_start()` to
`span_end() >= span_start()`, permitting empty spans at the trait level.
The translation layer enforces stricter constraints when needed for
identity
derivation. The `span_len()` method now uses `checked_sub` with an
explicit
panic to catch invariant violations rather than a debug assertion.

Changes also include:
- Type consistency fix in `distributed.rs` (compare `u64` to `u64`
  instead of cast)
- Clarified `PreHashedHasher` documentation to explain that it captures
  the
  last write ≥ 8 bytes from derived Hash impl, not the first
- Added regression test for `PreHashedHasher` to catch future changes to
  how `derive(Hash)` calls the Hasher trait
- Minor doc prose improvements in `gossip-scanner-runtime.md` and
  `fs-persistence-pipeline.md` to clarify error handling flow and crate
  layering

+++ crates/gossip-scanner-runtime/src/distributed.rs
@@ -3157,7 +3157,7 @@ where
     // → findings persistence → watermark commit.
     let captured_findings = capture_sink.take_captured_findings();
     let detected_count = capture_sink.detected_finding_count();
-    if detected_count as usize != captured_findings.len() {
+    if detected_count != captured_findings.len() as u64 {
         return Err(DistributedRuntimeError::Runtime(ScanRuntimeError::Driver(
             anyhow::anyhow!(
                 "finding counter ({detected_count}) diverged from captured payload count ({}); \
